build: create streamlit_toolkit module
